refactor (exception) : exception package 이동
- christmas.domain.exception -> christmas.exception 으로 이동
- exception은 domain 뿐만 아니라 전 영역에서 같이 사용하기 때문에 패키지 위치를 이동함
