+++ src/main/java/christmas/domain/plan/EventDate.java
@@ -1,6 +1,6 @@
 package christmas.domain.plan;
 
-import christmas.domain.exception.DateInputException;
+import christmas.exception.DateInputException;
 import java.time.DayOfWeek;
 import java.time.LocalDate;
 import java.util.Set;
