+++ src/test/java/christmas/domain/plan/EventDateTest.java
@@ -3,7 +3,7 @@ package christmas.domain.plan;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import christmas.domain.exception.DateInputException;
+import christmas.exception.DateInputException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.params.ParameterizedTest;
